+++ js/RTCRtpReceiver.js
@@ -3,9 +3,17 @@
  */
 module.exports = RTCRtpReceiver;
 
-function RTCRtpReceiver(data) {
+function RTCRtpReceiver(pc, data) {
 	data = data || {};
 
-	this._pc = data.pc;
-	this.track = data.track;
+	this._pc = pc;
+	this.track = pc.getOrCreateTrack(data.track);
 }
+
+RTCRtpReceiver.prototype.update = function ({ track }) {
+	if (track) {
+		this.track = this._pc.getOrCreateTrack(track);
+	} else {
+		this.track = null;
+	}
+};