+++ lib/WebRTC.xcframework/ios-arm64_armv7/WebRTC.framework/Headers/RTCCameraPreviewView.h
@@ -11,15 +11,15 @@
 #import <Foundation/Foundation.h>
 #import <UIKit/UIKit.h>
 
-#import <WebRTC/RTCMacros.h>
+#import "RTCMacros.h"
 
 @class AVCaptureSession;
 
 /** RTCCameraPreviewView is a view that renders local video from an
  *  AVCaptureSession.
  */
-RTC_EXPORT
-@interface RTCCameraPreviewView : UIView
+RTC_OBJC_EXPORT
+@interface RTC_OBJC_TYPE (RTCCameraPreviewView) : UIView
 
 /** The capture session being rendered in the view. Capture session
  *  is assigned to AVCaptureVideoPreviewLayer async in the same