+++ lib/WebRTC.xcframework/ios-arm64_armv7/WebRTC.framework/Headers/RTCSessionDescription.h
@@ -10,7 +10,7 @@
 
 #import <Foundation/Foundation.h>
 
-#import <WebRTC/RTCMacros.h>
+#import "RTCMacros.h"
 
 /**
  * Represents the session description type. This exposes the same types that are
@@ -20,12 +20,13 @@ typedef NS_ENUM(NSInteger, RTCSdpType) {
   RTCSdpTypeOffer,
   RTCSdpTypePrAnswer,
   RTCSdpTypeAnswer,
+  RTCSdpTypeRollback,
 };
 
 NS_ASSUME_NONNULL_BEGIN
 
-RTC_EXPORT
-@interface RTCSessionDescription : NSObject
+RTC_OBJC_EXPORT
+@interface RTC_OBJC_TYPE (RTCSessionDescription) : NSObject
 
 /** The type of session description. */
 @property(nonatomic, readonly) RTCSdpType type;