+++ lib/WebRTC.xcframework/ios-i386_x86_64-simulator/WebRTC.framework/Headers/RTCDataChannelConfiguration.h
@@ -11,12 +11,12 @@
 #import <AvailabilityMacros.h>
 #import <Foundation/Foundation.h>
 
-#import <WebRTC/RTCMacros.h>
+#import "RTCMacros.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
-RTC_EXPORT
-@interface RTCDataChannelConfiguration : NSObject
+RTC_OBJC_EXPORT
+@interface RTC_OBJC_TYPE (RTCDataChannelConfiguration) : NSObject
 
 /** Set to YES if ordered delivery is required. */
 @property(nonatomic, assign) BOOL isOrdered;