+++ lib/WebRTC.xcframework/ios-i386_x86_64-simulator/WebRTC.framework/Headers/RTCDispatcher.h
@@ -10,7 +10,7 @@
 
 #import <Foundation/Foundation.h>
 
-#import <WebRTC/RTCMacros.h>
+#import "RTCMacros.h"
 
 typedef NS_ENUM(NSInteger, RTCDispatcherQueueType) {
   // Main dispatcher queue.
@@ -20,13 +20,15 @@ typedef NS_ENUM(NSInteger, RTCDispatcherQueueType) {
   RTCDispatcherTypeCaptureSession,
   // Used for operations on AVAudioSession.
   RTCDispatcherTypeAudioSession,
+  // Used for operations on NWPathMonitor.
+  RTCDispatcherTypeNetworkMonitor,
 };
 
 /** Dispatcher that asynchronously dispatches blocks to a specific
  *  shared dispatch queue.
  */
-RTC_EXPORT
-@interface RTCDispatcher : NSObject
+RTC_OBJC_EXPORT
+@interface RTC_OBJC_TYPE (RTCDispatcher) : NSObject
 
 - (instancetype)init NS_UNAVAILABLE;