+++ lib/WebRTC.xcframework/ios-i386_x86_64-simulator/WebRTC.framework/Headers/RTCVideoCodecConstants.h
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2017 The WebRTC project authors. All Rights Reserved.
+ *  Copyright 2018 The WebRTC project authors. All Rights Reserved.
  *
  *  Use of this source code is governed by a BSD-style license
  *  that can be found in the LICENSE file in the root of the source
@@ -10,16 +10,7 @@
 
 #import <Foundation/Foundation.h>
 
-NS_ASSUME_NONNULL_BEGIN
+#import "RTCMacros.h"
 
-@interface RTCIntervalRange : NSObject
-
-@property(nonatomic, readonly) NSInteger min;
-@property(nonatomic, readonly) NSInteger max;
-
-- (instancetype)init;
-- (instancetype)initWithMin:(NSInteger)min max:(NSInteger)max NS_DESIGNATED_INITIALIZER;
-
-@end
-
-NS_ASSUME_NONNULL_END
+RTC_OBJC_EXPORT extern NSString* const kRTCVideoCodecVp8Name;
+RTC_OBJC_EXPORT extern NSString* const kRTCVideoCodecVp9Name;