+++ lib/WebRTC.xcframework/ios-i386_x86_64-simulator/WebRTC.framework/Headers/RTCVideoDecoderVP9.h
@@ -10,16 +10,16 @@
 
 #import <Foundation/Foundation.h>
 
-#import <WebRTC/RTCMacros.h>
-#import <WebRTC/RTCVideoCodec.h>
+#import "RTCMacros.h"
+#import "RTCVideoDecoder.h"
 
-RTC_EXPORT
-@interface RTCVideoDecoderVP9 : NSObject
+RTC_OBJC_EXPORT
+@interface RTC_OBJC_TYPE (RTCVideoDecoderVP9) : NSObject
 
 /* This returns a VP9 decoder that can be returned from a RTCVideoDecoderFactory injected into
  * RTCPeerConnectionFactory. Even though it implements the RTCVideoDecoder protocol, it can not be
  * used independently from the RTCPeerConnectionFactory.
  */
-+ (id<RTCVideoDecoder>)vp9Decoder;
++ (id<RTC_OBJC_TYPE(RTCVideoDecoder)>)vp9Decoder;
 
 @end